% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_method.R
\name{assign_method}
\alias{assign_method}
\title{Assign volume esitmation method in to each individuals}
\usage{
assign_method(data, method_file = NULL)
}
\arguments{
\item{data}{The data that records observations of each individuals size measurements
with at least the following columns: "Taxon"}

\item{method_file}{The file that contains the pairwise. If NULL, the method file will use the default
"biovolume_method" data.}
}
\value{

}
\description{
Assign volume esitmation method in to each individuals
}
\examples{
a <- data.frame(Taxon = c("Polychaeta", "Oligochaeta", "Sipuncula", "Not in the column"))
assign_method(a)
}
